More run 3 updates

+++ campaigns/Run3Summer23BPixwmLHE/getpileupfiles.sh
@@ -2,14 +2,15 @@
 
 if [ -z "$1" ]; then
     CERNNAME=$(whoami)
-    echo "Using current username (${CERNNAME}) as CERN username"
-    echo " "
-    echo " "
-    CERNNAME=$(whoami)
 else
     CERNNAME=$1
 fi
+echo "[${0}] Using {CERNNAME} as username for Rucio"
 
-mv pileupinput.dat pileupinput.dat.$(date +"%Y-%m-%d-%H-%M-%S")
+PILEUP_DATASET="/Neutrino_E-10_gun/Run3Summer21PrePremix-Summer23BPix_130X_mcRun3_2023_realistic_postBPix_v1-v1/PREMIX"
+
+if [ -f pileupinput.dat ]; then
+    mv pileupinput.dat pileupinput.dat.$(date +"%Y-%m-%d-%H-%M-%S")
+fi
 # dasgoclient -query="file dataset=/Neutrino_E-10_gun/Run3Summer21PrePremix-Summer22_124X_mcRun3_2022_realistic_v11-v2/PREMIX" > pileupinput.dat
-python3 ../../utils/get_files_on_disk.py -u ${CERNNAME} -o pileupinput.dat /Neutrino_E-10_gun/Run3Summer21PrePremix-Summer23BPix_130X_mcRun3_2023_realistic_postBPix_v1-v1/PREMIX
+get_files_on_disk.py -u ${CERNNAME} -o pileupinput.dat ${PILEUP_DATASET}
