+++ campaigns/RunIISummer20UL17wmLHE/getpileupfiles.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
-mv pileupinput.dat pileupinput.dat.$(date +"%Y-%m-%d-%H-%M-%S")
+if [ -f pileupinput.dat ]; then
+	mv pileupinput.dat pileupinput.dat.$(date +"%Y-%m-%d-%H-%M-%S")
+fi
 dasgoclient -query="file dataset=/Neutrino_E-10_gun/RunIISummer20ULPrePremix-UL17_106X_mc2017_realistic_v6-v3/PREMIX" > pileupinput.dat
